poc idea for golang statemachine code generation - pretty much a hacked copy of generate_cpp.py.
This is my first attempt at golang and no idea if the state machine ideas will pan out.
I like the interface mechanism of golang.
Wish I could figure out how to make a pointer receiver const.
Not sure how this will ultimately be packaged as well. The engines will need to be separated from the generators.
Will be cleaner as well.
Maybe I should have kept the scratchpad name for the repo.

The copy/paste of generate_cpp.py and generate_go.py and the use of print will have to be resolved.